refactor(amf-deserializer): more refactor
